Time: 9 ms (97.97%), Space: 47.4 MB (16.16%) - LeetHub

+++ 0462-minimum-moves-to-equal-array-elements-ii/0462-minimum-moves-to-equal-array-elements-ii.java
@@ -3,12 +3,12 @@ class Solution {
     {
         Arrays.sort(nums);
         int median=nums[nums.length/2];
-        long moves=0;
+        int moves=0;
 
         for(int num:nums){
             moves+=Math.abs(num-median);
         }
-        return (int) moves;
+        return moves;
 
     }
 }
